refactor: 사용자 위치 버튼에 대한 함수를 별도의 Extension으로 코드를 분리
