+++ PassportTrails/Scenes/StampMapScene/StampMapViewController.swift
@@ -20,7 +20,6 @@ final class StampMapViewController: BaseViewController {
         
         configureLocationManager()
         configureMapView()
-        configureUserTrackingButton()
         
         fetchGeoJson(fileName: "place")
     }
@@ -73,46 +72,26 @@ final class StampMapViewController: BaseViewController {
             mapView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             mapView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             mapView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            mapView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            mapView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
         ])
     }
     
     private func configureMapView() {
         mapView.delegate = self
         mapView.register(PlaceAnnotationView.self, forAnnotationViewWithReuseIdentifier: MKMapViewDefaultAnnotationViewReuseIdentifier)
+        
         mapView.mapType = .standard
         mapView.showsUserLocation = true
         mapView.showsCompass = true
         mapView.showsScale = true
+        
+        mapView.configureUserTrackingButton()
     }
     
     private func configureLocationManager() {
         locationManager.delegate = self
         
         checkDeviceLocationAuthorization()
-    }
-    
-    private func configureUserTrackingButton() {
-        let trackingButton: MKUserTrackingBarButtonItem = MKUserTrackingBarButtonItem(mapView: mapView)
-        trackingButton.customView?.tintColor = .black
-        trackingButton.customView?.frame.size = CGSize(width: 50, height: 50)
-        
-        let toolBarFrame = CGRect(origin: .zero, size: CGSize(width: 50, height: 50))
-        let toolbar = UIToolbar(frame: toolBarFrame)
-        toolbar.barTintColor = UIColor.white
-        toolbar.isTranslucent = false
-        
-        let flex: UIBarButtonItem = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: self, action: nil)
-        toolbar.items = [flex, trackingButton, flex]
-        
-        let origin = CGPoint(x: self.view.frame.size.width - 85, y: 60)
-        let roundedSquare = UIView(frame: CGRect(origin: origin, size: CGSize(width: 50, height: 50)))
-        roundedSquare.backgroundColor = UIColor.white
-        roundedSquare.layer.cornerRadius = 5
-        roundedSquare.layer.masksToBounds = true
-        
-        mapView.addSubview(roundedSquare)
-        roundedSquare.addSubview(toolbar)
     }
     
     //MARK: LocationAuthorization
